feat: The `attachment` submodule has been updated to take in a list of scope IDs<br>- The `scope` input has been replaced by `scope_ids`. Scope is no longer created by the `attachment` submodule, as it is now a first class citizen and has its own IBM terraform provider resource `ibm_scc_scope` (#246)
BREAKING CHANGE: Upgrading to this version is going to destroy and recreate any existing attachment that was created using the `attachment` submodule. Please note that the scope used in the attachment that is being destroyed is not going to be deleted by terraform since previously scope was not managed by terraform state. Manual cleanup may be required if you want to remove the scope.